refactor: split lib/db/db.mjs into focused domain modules
db.mjs had grown to 5300+ lines covering connection lifecycle,
migrations, stats, activity state, semantic memory writes, episode
search, prompt context assembly, and more, all in one LoreDb class.
Split each cohesive group of methods into its own db-*.mjs module
(activity-state, stats, scope-management, semantic-memory-write,
episode-search, prompt-context, etc.), following the existing
db-migration-runner.mjs / db-retrieval-policy.mjs pattern already
used in this directory. LoreDb keeps its full public API via thin
delegator methods; no behavior changed. All 185 prototype methods
and the full test suite (1286 tests) are unchanged and passing.
