+++ lib/db/db-improvement-artifacts.mjs
@@ -379,3 +379,193 @@ export function buildBackfillRunSummaryUpdateImpl(db, runId, { lastError = null 
     lastError: derivedLastError,
   };
 }
+
+export function upsertImprovementArtifact(owner, {
+  sourceCaseId,
+  sourceKind,
+  title,
+  summary,
+  evidence = {},
+  trace = {},
+  linkedMemoryId = null,
+  repository = null,
+}) {
+  owner.ensureOpen();
+  return upsertImprovementArtifactImpl(owner.db, {
+    sourceCaseId,
+    sourceKind,
+    title,
+    summary,
+    evidence,
+    trace,
+    linkedMemoryId,
+    repository,
+  });
+}
+
+export function updateImprovementArtifactStatus(owner, {
+  id,
+  status,
+  supersededBy = null,
+}) {
+  owner.ensureOpen();
+  const normalizedId = String(id || "").trim();
+  if (!normalizedId) {
+    throw new Error("id is required");
+  }
+  const nextStatus = normalizeImprovementStatus(status, IMPROVEMENT_STATUS.ACTIVE);
+  const timestamp = nowIso();
+  const resolvedAt = nextStatus === IMPROVEMENT_STATUS.RESOLVED ? timestamp : null;
+  owner.db.prepare(`
+    UPDATE improvement_backlog
+    SET status = ?,
+        superseded_by = CASE
+          WHEN ? = 'superseded' THEN COALESCE(?, superseded_by)
+          ELSE NULL
+        END,
+        review_state = CASE
+          WHEN ? = 'superseded' AND proposal_path IS NOT NULL THEN 'superseded'
+          ELSE review_state
+        END,
+        resolved_at = ?,
+        updated_at = ?
+    WHERE id = ?
+  `).run(
+    nextStatus,
+    nextStatus,
+    supersededBy ?? null,
+    nextStatus,
+    resolvedAt,
+    timestamp,
+    normalizedId,
+  );
+}
+
+export function getImprovementArtifact(owner, id) {
+  owner.ensureOpen();
+  const normalizedId = String(id || "").trim();
+  if (!normalizedId) {
+    return null;
+  }
+  const row = owner.db.prepare(`
+    SELECT
+      id,
+      source_case_id,
+      source_kind,
+      title,
+      summary,
+      evidence_json,
+      trace_json,
+      status,
+      linked_memory_id,
+      repository,
+      superseded_by,
+      proposal_type,
+      proposal_path,
+      proposal_hash,
+      review_state,
+      review_requested_at,
+      review_requested_by,
+      reviewer_decision,
+      reviewer_notes_json,
+      created_at,
+      updated_at,
+      resolved_at
+    FROM improvement_backlog
+    WHERE id = ?
+    LIMIT 1
+  `).get(normalizedId);
+  if (!row) {
+    return null;
+  }
+  return {
+    ...row,
+    evidence: parseJsonObject(row.evidence_json),
+    trace: parseJsonObject(row.trace_json),
+    reviewer_notes: parseJsonObject(row.reviewer_notes_json),
+  };
+}
+
+export function setImprovementArtifactProposal(owner, {
+  id,
+  proposalType,
+  proposalPath,
+  proposalHash,
+  reviewState = IMPROVEMENT_REVIEW_STATE.DRAFT,
+  reviewRequestedAt = null,
+  reviewRequestedBy = null,
+  reviewerDecision = null,
+  reviewerNotes = {},
+}) {
+  owner.ensureOpen();
+  setImprovementArtifactProposalImpl(owner.db, {
+    id,
+    proposalType,
+    proposalPath,
+    proposalHash,
+    reviewState,
+    reviewRequestedAt,
+    reviewRequestedBy,
+    reviewerDecision,
+    reviewerNotes,
+  });
+}
+
+export function listImprovementArtifacts(owner, {
+  sourceKind,
+  sourceCaseId,
+  status,
+  reviewState,
+  hasProposal,
+  updatedBefore,
+  repository,
+  sort = "updated_desc",
+  limit = 10,
+} = {}) {
+  owner.ensureOpen();
+  const { where, params } = buildImprovementArtifactFilters({
+    sourceKind,
+    sourceCaseId,
+    status,
+    reviewState,
+    hasProposal,
+    updatedBefore,
+    repository,
+  });
+  params.push(limit);
+  const rows = owner.db.prepare(`
+    SELECT
+      id,
+      source_case_id,
+      source_kind,
+      title,
+      summary,
+      evidence_json,
+      trace_json,
+      status,
+      linked_memory_id,
+      repository,
+      superseded_by,
+      proposal_type,
+      proposal_path,
+      proposal_hash,
+      review_state,
+      review_requested_at,
+      review_requested_by,
+      reviewer_decision,
+      reviewer_notes_json,
+      created_at,
+      updated_at,
+      resolved_at
+    FROM improvement_backlog
+    ${where.length > 0 ? `WHERE ${where.join(" AND ")}` : ""}
+    ORDER BY updated_at ${sort === "updated_asc" ? "ASC" : "DESC"}
+    LIMIT ?
+  `).all(...params);
+  return rows.map((row) => ({
+    ...row,
+    evidence: parseJsonObject(row.evidence_json),
+    trace: parseJsonObject(row.trace_json),
+    reviewer_notes: parseJsonObject(row.reviewer_notes_json),
+  }));
+}
