+++ lib/db/db-trace-samples.mjs
@@ -196,3 +196,137 @@ export function listRetrievalTraceSamplesImpl(db, {
     recordedAt: row.recorded_at,
   }));
 }
+
+export function normalizeRetrievalTraceHook(owner, hook) {
+  return String(hook || "").trim();
+}
+
+export function normalizeRetrievalTraceScopeType(owner, scopeType) {
+  return scopeType === "global" ? "global" : "repo";
+}
+
+export function normalizeRetrievalTraceSectionTitles(owner, sectionTitles) {
+  if (!Array.isArray(sectionTitles)) {
+    return [];
+  }
+  return sectionTitles
+    .slice(0, 8)
+    .map((title) => (title == null ? "" : String(title)));
+}
+
+export function buildRetrievalTraceSampleRecord(owner, {
+  id = null,
+  repository = null,
+  scopeType = "repo",
+  hook,
+  route = null,
+  routeReason = null,
+  contextInjected = false,
+  latencyMs = null,
+  promptPreview = "",
+  sectionTitles = [],
+  promptNeed = {},
+  eligibility = {},
+  lookups = {},
+  omissions = [],
+  output = {},
+  trace = {},
+  recordedAt = nowIso(),
+}) {
+  return buildRetrievalTraceSampleRecordImpl(owner, {
+    id,
+    repository,
+    scopeType,
+    hook,
+    route,
+    routeReason,
+    contextInjected,
+    latencyMs,
+    promptPreview,
+    sectionTitles,
+    promptNeed,
+    eligibility,
+    lookups,
+    omissions,
+    output,
+    trace,
+    recordedAt,
+  });
+}
+
+export function writeRetrievalTraceSample(owner, record) {
+  owner.db.prepare(`
+    INSERT INTO retrieval_trace_sample (
+      id,
+      repository,
+      scope_type,
+      hook,
+      route,
+      route_reason,
+      context_injected,
+      latency_ms,
+      prompt_preview,
+      section_titles_json,
+      prompt_need_json,
+      eligibility_json,
+      lookups_json,
+      omissions_json,
+      output_json,
+      trace_json,
+      recorded_at
+    ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+  `).run(
+    record.id,
+    record.repository,
+    record.scopeType,
+    record.hook,
+    record.route,
+    record.routeReason,
+    record.contextInjected,
+    record.latencyMs,
+    record.promptPreview,
+    JSON.stringify(record.sectionTitles),
+    record.promptNeed,
+    record.eligibility,
+    record.lookups,
+    record.omissions,
+    record.output,
+    record.trace,
+    record.recordedAt,
+  );
+}
+
+export function insertRetrievalTraceSample(owner, sample = {}) {
+  owner.ensureOpen();
+  const record = owner.buildRetrievalTraceSampleRecord(sample);
+  owner.writeRetrievalTraceSample(record);
+  return record.id;
+}
+
+export function pruneRetrievalTraceSamples(owner, {
+  repository = null,
+  maxRowsPerRepository = 120,
+  maxRowsGlobal = 240,
+  maxAgeMs = 14 * 24 * 60 * 60 * 1000,
+} = {}) {
+  owner.ensureOpen();
+  return pruneRetrievalTraceSamplesImpl(owner.db, {
+    repository,
+    maxRowsPerRepository,
+    maxRowsGlobal,
+    maxAgeMs,
+  });
+}
+
+export function listRetrievalTraceSamples(owner, {
+  repository,
+  includeGlobal = true,
+  limit = 10,
+} = {}) {
+  owner.ensureOpen();
+  return listRetrievalTraceSamplesImpl(owner.db, {
+    repository,
+    includeGlobal,
+    limit,
+  });
+}
